refactor(speech): consolidate STT providers
